gl_sheets_one_page/requirements.txt
psycopg2
django
djangorestframework
django-cors-headers
